+++ llama-index-core/llama_index/core/readers/file/base.py
@@ -77,14 +77,19 @@ def _format_file_timestamp(timestamp: float) -> Optional[str]:
 def default_file_metadata_func(
     file_path: str, fs: Optional[fsspec.AbstractFileSystem] = None
 ) -> Dict:
-    """Get some handy metadate from filesystem.
+    """Get some handy metadata from filesystem.
 
     Args:
         file_path: str: file path in str
     """
     fs = fs or get_default_fs()
     stat_result = fs.stat(file_path)
-    file_name = os.path.basename(str(stat_result["name"]))
+
+    try:
+        file_name = os.path.basename(str(stat_result["name"]))
+    except Exception as e:
+        file_name = os.path.basename(file_path)
+
     creation_date = _format_file_timestamp(stat_result.get("created"))
     last_modified_date = _format_file_timestamp(stat_result.get("mtime"))
     last_accessed_date = _format_file_timestamp(stat_result.get("atime"))