+++ llama-index-integrations/agent/llama-index-agent-openai-legacy/llama_index/agent/openai_legacy/openai_agent.py
@@ -248,7 +248,9 @@ class BaseOpenAIAgent(BaseAgent):
             chat_stream_response.awrite_response_to_history(self.memory)
         )
         # wait until openAI functions stop executing
+        chat_stream_response._ensure_async_setup()
         await chat_stream_response._is_function_false_event.wait()
+
         # return response stream
         return chat_stream_response