+++ llama-index-integrations/readers/llama-index-readers-file/llama_index/readers/file/docs/base.py
@@ -53,7 +53,7 @@ class PDFReader(BaseReader):
             # This block returns a whole PDF as a single Document
             if self.return_full_document:
                 text = ""
-                metadata = {"file_name": fp.name}
+                metadata = {"file_name": file.name}
 
                 for page in range(num_pages):
                     # Extract the text from the page
@@ -71,7 +71,7 @@ class PDFReader(BaseReader):
                     page_text = pdf.pages[page].extract_text()
                     page_label = pdf.page_labels[page]
 
-                    metadata = {"page_label": page_label, "file_name": fp.name}
+                    metadata = {"page_label": page_label, "file_name": file.name}
                     if extra_info is not None:
                         metadata.update(extra_info)